requirements.txt
# Real-Time Data Pipeline - Python Dependencies
# Version: 2.0.0
# Python: 3.8+

# Core data processing
pyspark==3.4.1

# Kafka integration
kafka-python==2.0.2

# HTTP requests for API calls
requests==2.31.0

# Data manipulation and CSV export
pandas==2.1.4

# Optional: Enhanced data analysis
# numpy==1.24.3
# pyarrow==14.0.1  # For optimized Parquet operations

# Development dependencies (uncomment if needed)
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0 